customize docker files to build prod image
