+++ docker-compose.yml
@@ -16,76 +16,34 @@
 #
 x-superset-build: &superset-build
   args:
-    NPM_BUILD_CMD: build-dev
+    NPM_BUILD_CMD: build
   context: ./
   dockerfile: Dockerfile
-  target: dev
-x-superset-depends-on: &superset-depends-on
-  - db
-  - redis
-x-superset-volumes: &superset-volumes
-  # /app/pythonpath_docker will be appended to the PYTHONPATH in the final container
-  - ./docker/docker-init.sh:/app/docker-init.sh
-  - ./docker/pythonpath_dev:/app/pythonpath
-  - ./superset:/app/superset
-  - ./superset-frontend:/app/superset-frontend
-  - superset_home:/app/superset_home
+  target: prod
 
 version: "3.7"
 services:
-  redis:
-    image: redis:3.2
-    restart: unless-stopped
-    ports:
-      - "127.0.0.1:6379:6379"
-    volumes:
-      - redis:/data
-
-  db:
-    env_file: docker/.env
-    image: postgres:10
-    restart: unless-stopped
-    ports:
-      - "127.0.0.1:5432:5432"
-    volumes:
-      - db_home:/var/lib/postgresql/data
 
   superset:
+    image: 'dwh:superset-fork'
+    container_name: dwh-superset-fork
     build: *superset-build
     command: ["flask", "run", "-p", "8088", "--with-threads", "--reload", "--debugger", "--host=0.0.0.0"]
     env_file: docker/.env
     restart: unless-stopped
     ports:
       - 8088:8088
-    depends_on: *superset-depends-on
-    volumes: *superset-volumes
+    environment:
+      - SUPERSET_CONFIG_PATH=/app/superset_config.py
 
   superset-init:
+    image: 'dwh:superset-init'
+    container_name: dwh-superset-init
     build: *superset-build
     command: ["/app/docker-init.sh"]
     env_file: docker/.env
-    depends_on: *superset-depends-on
-    volumes: *superset-volumes
 
-  superset-node:
-    image: node:10-jessie
-    command: ["bash", "-c", "cd /app/superset-frontend && npm install --global webpack webpack-cli && npm install && npm run dev"]
-    env_file: docker/.env
-    depends_on: *superset-depends-on
-    volumes: *superset-volumes
-
-  superset-worker:
-    build: *superset-build
-    command: ["celery", "worker", "--app=superset.tasks.celery_app:app", "-Ofair"]
-    env_file: docker/.env
-    restart: unless-stopped
-    depends_on: *superset-depends-on
-    volumes: *superset-volumes
-
-volumes:
-  superset_home:
-    external: false
-  db_home:
-    external: false
-  redis:
-    external: false
+networks:
+  default:
+    external:
+      name: superset_backend
